+++ fpdf/fpdf.py
@@ -33,12 +33,14 @@ from enum import IntEnum
 from functools import wraps
 from pathlib import Path
 from typing import Callable, NamedTuple, Optional, Union, List
+from xml.etree.ElementTree import ParseError, XML
 
 from PIL import Image
 
 from .actions import Action
 from .errors import FPDFException, FPDFPageFormatException
 from .fonts import fpdf_charwidths
+from .graphics_state import GraphicsStateMixin
 from .image_parsing import get_img_info, load_image, SUPPORTED_IMAGE_FILTERS
 from .line_break import Fragment, MultiLineBreak
 from .outline import serialize_outline, OutlineSection
@@ -46,7 +48,7 @@ from . import drawing
 from .recorder import FPDFRecorder
 from .structure_tree import MarkedContent, StructureTreeBuilder
 from .ttfonts import TTFontFile
-from .graphics_state import GraphicsStateMixin
+from .svg import Percent, SVGObject
 from .util import (
     enclose_in_parens,
     escape_parens,
@@ -2724,11 +2726,17 @@ class FPDF(GraphicsStateMixin):
                 '"type" is unused and will soon be deprecated',
                 PendingDeprecationWarning,
             )
+        if str(name).endswith(".svg"):
+            # Insert it as a PDF path:
+            img = load_image(str(name))
+            return self._vector_image(img, x, y, w, h, link, title, alt_text)
         if isinstance(name, str):
             img = None
         elif isinstance(name, Image.Image):
             name, img = hashlib.md5(name.tobytes()).hexdigest(), name
         elif isinstance(name, io.BytesIO):
+            if _is_xml(name):
+                return self._vector_image(name, x, y, w, h, link, title, alt_text)
             name, img = hashlib.md5(name.getvalue()).hexdigest(), name
         else:
             name, img = str(name), name
@@ -2760,7 +2768,6 @@ class FPDF(GraphicsStateMixin):
             self._perform_page_break_if_need_be(h)
             y = self.y
             self.y += h
-
         if x is None:
             x = self.x
 
@@ -2777,6 +2784,66 @@ class FPDF(GraphicsStateMixin):
             self.link(x, y, w, h, link)
 
         return info
+
+    def _vector_image(
+        self,
+        img: io.BytesIO,
+        x=None,
+        y=None,
+        w=0,
+        h=0,
+        link="",
+        title=None,
+        alt_text=None,
+    ):
+        svg = SVGObject(img.getvalue())
+        if w == 0 and h == 0:
+            if not svg.width or not svg.height:
+                raise ValueError(
+                    '<svg> has no "height" / "width": w= or h= must be provided to FPDF.image()'
+                )
+            w = (
+                svg.width * self.epw / 100
+                if isinstance(svg.width, Percent)
+                else svg.width
+            )
+            h = (
+                svg.height * self.eph / 100
+                if isinstance(svg.height, Percent)
+                else svg.height
+            )
+        else:
+            _, _, vw, vh = svg.viewbox
+            if w == 0:
+                w = vw * h / vh
+            elif h == 0:
+                h = vh * w / vw
+
+        # Flowing mode
+        if y is None:
+            self._perform_page_break_if_need_be(h)
+            y = self.y
+            self.y += h
+        if x is None:
+            x = self.x
+
+        _, _, path = svg.transform_to_rect_viewport(
+            scale=1, width=w, height=h, ignore_svg_top_attrs=True
+        )
+        path.transform = path.transform @ drawing.Transform.translation(x, y)
+
+        try:
+            old_x, old_y = self.x, self.y
+            self.set_xy(0, 0)
+            if title or alt_text:
+                with self._marked_sequence(title=title, alt_text=alt_text):
+                    self.draw_path(path)
+            else:
+                self.draw_path(path)
+        finally:
+            self.set_xy(old_x, old_y)
+        if link:
+            self.link(x, y, w, h, link)
 
     def _downscale_image(self, name, img, info, w, h):
         width_in_pt, height_in_pt = w * self.k, h * self.k
@@ -4114,6 +4181,14 @@ def _sizeof_fmt(num, suffix="B"):
             return f"{num:3.1f}{unit}{suffix}"
         num /= 1024
     return f"{num:.1f}Yi{suffix}"
+
+
+def _is_xml(img: io.BytesIO):
+    try:
+        XML(img.getvalue())
+        return True
+    except ParseError:
+        return False
 
 
 sys.modules[__name__].__class__ = WarnOnDeprecatedModuleAttributes